wip econfig

+++ server/src/electronConfig.ts
@@ -55,7 +55,9 @@ export const getElectronConfig = () => {
     getCredentials,
     setCredentials: (connection?: DBSConnectionInfo) => {
       if(!connection){
-        fs.unlinkSync(electronConfigPath);
+        if(fs.existsSync(electronConfigPath)){
+          fs.unlinkSync(electronConfigPath);
+        }
       } else {
         fs.writeFileSync(electronConfigPath, safeStorage!.encryptString(JSON.stringify(connection)));
       }
